Multi-Agent-project/Manual-Agents/requirements.txt
geopy
plyer
python-dotenv
pytz
requests
streamlit
yfinance
